+++ Nicuga/Game.h
@@ -52,17 +52,17 @@ public:
 	int getScore();
 	Level * getCurrentLevel();
 	int getCurrentLevelPosition();
-	void setOnDestroyCallback(void (*onDestroyCallback)(void * userdata,unsigned int bulletId,Ship * ship, HostileInstance * hostile,float x,float y),void * userdata);
-	void setOnHitCallback(void (*onHitCallback)(void * userdata,unsigned int bulletId,Ship * ship, HostileInstance * hostile,float x,float y),void * userdata);
+	void setOnDestroyCallback(void (*onDestroyCallback)(void * userdata,unsigned int bulletId,Ship * ship, HostileInstance * hostile,short x,short y),void * userdata);
+	void setOnHitCallback(void (*onHitCallback)(void * userdata,unsigned int bulletId,Ship * ship, HostileInstance * hostile,short x,short y),void * userdata);
 	int getRemainingPlayerLife();
 	bool isOver();
 	void respawnShip();
 	void setPlayerLives(int lives);
 private:
-	void (*onDestroyCallback)(void * userdata,unsigned int bulletId,Ship * ship, HostileInstance * hostile,float x,float y);
+	void (*onDestroyCallback)(void * userdata,unsigned int bulletId,Ship * ship, HostileInstance * hostile,short x,short y);
 	void * onDestroyCallbackUserData;
 
-	void (*onHitCallback)(void * userdata,unsigned int bulletId,Ship * ship, HostileInstance * hostile,float x,float y);
+	void (*onHitCallback)(void * userdata,unsigned int bulletId,Ship * ship, HostileInstance * hostile,short x,short y);
 	void * onHitCallbackUserData;
 	
 	Level * currentLevel;
